Updating Chapt_01 .vcxproj

+++ FENGYUANG/Chapt_01/Hello3/Hello3.cpp
@@ -40,7 +40,7 @@ void CenterText(HDC hDC, int x, int y, LPCTSTR szFace, LPCTSTR szMessage, int po
 
     SetBkMode(hDC, TRANSPARENT);
     SetTextColor(hDC, RGB(0, 0, 0xFF));
-    TextOut(hDC, x, y, szMessage, _tcslen(szMessage));
+    TextOut(hDC, x, y, szMessage, (int)_tcslen(szMessage));
 
     SelectObject(hDC, hOld);
     DeleteObject(hFont);
@@ -83,5 +83,5 @@ int WINAPI WinMain(HINSTANCE hInst, HINSTANCE,
 
     win.UpdateWindow();
   
-    return win.MessageLoop();
+    return (int)win.MessageLoop();
 }
